[#94] spring graphql 컨트롤러 임시적용
