+++ app/src/main/resources/application.yml
@@ -1,7 +1,3 @@
-graphql:
-  packages:
-    - "com.santaclose.app"
-
 spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
